Enable building with Bazel 8
Recent versions of abseil strongly encourage using
Bazel modules. Therefore, switch some dependencies
over to using modules as well, but leave a few
using the old WORKSPACE syntax for now. To enable
this, we rename WORKSPACE to WORKSPACE.bzlmod
(Bazel calls this "hybrid mode").

This also involves upgrading the versions of
several of our dependencies, notably protobuf and
abseil. It would be nice to defer doing this for
now, but some initial investigation suggested that
this would be quite difficult due to the web of
dependencies.