Add an empty WORKSPACE file
This appears to be necessary to workaround a bug or odd behavior in
recent versions of Bazel: in eventuals, submodules/stout was considered
to be part of the eventuals repo (despite being declared as a
local_repository) because of the renaming of WORKSPACE.bazel to
WORKSPACE.bzlmod. For the time being, re-add an empty WORKSPACE file.